Implement dependency-id
Generate unique dependency-id. The dep-id will
also allow installing multiple library versions
at the same time. Developers can set a different
dependency-id if the break ABI, for example if
the id is "libfoo-1" with version number "1.2.4",
and the ABI is broken in the next release, they can
just set the new dep-id "libfoo-2", for example.