updater: Get rid of SoftwareItem
We now use the GObject type system to determine the type of the
submitted software update.
By doing this, all information in AppItem and IPKDependency objects will
remain intact and frontends can make use of this information to show user-
friendly updater dialogs.
(SoftwareItem was introduced with the last commit, so this is not
exactly an API break)