tests: Port to new testing infrastructure
This will make it possible to get rid of the misuse of InstallMode.TEST
later and to improve testing in general. (Planned for 0.5.8)
The basic migration done with this patch actually already spotted a new
bug which appears when resolving .desktop-files in certain corner cases.
The bug is fixed with this commit.