plugin: Fix interaction with PackageKit messages
Apparently the whole communication with PackageKit had some serious
issues due to old code from the pre-PkBackendJob age. This resulted in a
large amount of bugs, appearing due to false communication between
PackageKit and Listaller.
This patch resolves all issues associated with that. (And cleans up the
plugin code a little, so nobody will hit this bug again.)