lilith: Initialize
Welcome Lilith, the Listaller IPK package lint!
This tool will later be able to perform a series of checks on Listaller
IPK packages, to prevent "bad" packages to be made available in the
wild.
Lilith may also be used to check Listaller package sources.
While first Lilith was planned to be written in Python or Perl to make
it easily extendable, Lilith checks will now be written in Vala. This
makes it possible to use already existing Listaller facilities for
package validation.
However, to write Listaller checks, only little Vala knowledge will be
required, as Lilith will provide many preformed functions which make
writing and extending checks as easy as possible.