spec: New control value to specify if an user has to accept a license
If the IPK control value "UserAcceptLicense" is set to "yes", the user
will be prompted to accept the supplied license at install time.
If this value is not set, "no" is assumed. (And the user is not bothered
with pointless EULAs or attempts to accept the GPL over and over again)