Merge parts of the item-facility from DepInfo
Merge some parts of the item-checkup facilities from the DepInfo object,
which should soon be replaced with the new system.
The Dep.Module class might later receive the ItemInstall stuff.

Please note that the previous "component" naming has changed to "item",
which is much more descriptive, and also doesn't conflict with the new,
bigger meaning of a component.